chore($ui/detail): move package
