+++ app/src/main/java/woowacourse/movie/ui/detail/view/ScreenDetailView.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui.detail
+package woowacourse.movie.ui.detail.view
 
 import woowacourse.movie.ui.ScreenDetailUI
